refactor: use clap for argument parsing
