--- src/cli/mod.rs
@@ -1,2 +1,0 @@
-pub mod flag_registry;
-pub mod cli_interface;
